+++ src/lib/Scrolly/Scrolly.docs.mdx
@@ -1,0 +1,37 @@
+import * as Stories from "./Scrolly.stories.svelte"
+import { Meta, Controls, ArgTypes, Markdown, Story } from "@storybook/blocks"
+
+<Meta component={Stories.Scrolly} of={Stories} />
+
+# Scrolly
+
+A component that can be used to create "scrolly" sections and pages. This component is a wrapper around the [svelte-scroller](https://github.com/sveltejs/svelte-scroller) library. If you're project has needs that go beyond what this component can support, you should consider using `svelte-scroller` directly. But for many common scenarious, the options provided here should have you covered.
+
+Basic usage:
+```html
+<script>
+  import { Scrolly } from "@urban-institute/dataviz-components";
+
+  const slides = ["Slide 1", "Slide 2", "Slide 3"];
+
+</script>
+<Scrolly slides={slides}>
+  <div
+    slot="background"
+    style="width: 100%; height: 100vh; color: var(--color-white); font-weight: var(--font-weight-bold); background: var(--color-blue); display: flex; align-items: center; justify-content: center;"
+  >
+    Scrolly background
+  </div>
+</Scrolly>
+````
+
+<Story of={Stories.Default} />
+<Controls />
+
+## Creating custom backgrounds
+
+Custom background information TK.
+
+## Custom foregrounds
+
+Information about overriding the default foreground elements TK.
