+++ src/lib/index.js
@@ -20,3 +20,4 @@ export { default as FontsUrban } from "./Fonts/FontsUrban.svelte";
 export { default as FontsWorkrise } from "./Fonts/FontsWorkrise.svelte";
 export { default as Meta } from "./Meta/Meta.svelte";
 export { default as Toggle } from "./Toggle/Toggle.svelte";
+export { default as Scroller } from "./Scroller/Scroller.svelte";
